+++ src/main/resources/sql/tsql/createStoreLong.sql
@@ -14,7 +14,7 @@
 CREATE TABLE JdsStoreLong(
 	FieldId     BIGINT,
 	EntityGuid  NVARCHAR(48) NOT NULL,
-	Value       INTEGER,
+	Value       BIGINT,
 	PRIMARY KEY (FieldId,EntityGuid),
 	CONSTRAINT fk_JdsStoreLong_ParentEntityGuid FOREIGN KEY (EntityGuid) REFERENCES JdsStoreEntityOverview(EntityGuid) ON DELETE CASCADE
 );
